refactor: collocate errors with `Serializer`/`Deserializer` accordingly
